separate chain token list logic, set reset tx states on navigate to history
