+++ src/app/bridge/_components/bridge-review/TransactionStepper.tsx
@@ -29,7 +29,14 @@ export default function BridgeTransactionStepper({ steps, onCloseModal, currentS
       </DialogClose>
       <div className="flex flex-col md:flex-row items-center md:items-start justify-center md:gap-x-16 gap-y-16 py-5">
         {steps.map((step, index) => (
-          <TransactionStep key={index} currentStep={currentStep} index={index} step={step} steps={steps} />
+          <TransactionStep
+            key={index}
+            onResetTransaction={resetTransaction}
+            currentStep={currentStep}
+            index={index}
+            step={step}
+            steps={steps}
+          />
         ))}
       </div>
       <Stepper totalSteps={steps.length} currentStep={currentStep.count} />
