+++ src/app/bridge/_components/select-token/_logic/index.tsx
@@ -248,6 +248,7 @@ const SelectTokenLogic = () => {
     setShowWalletAddress,
     showWalletAddress,
     actionButtonStatus,
+    isWalletConnected,
   };
 };
